Refactor polarization module location
- Moved the polarization module to a new directory structure.
- Updated imports across multiple files to reflect the new path.
